+++ Test/CoreSDK.Test/Shared/Extensibility/TelemetryConfigurationFactoryTest.cs
@@ -598,7 +598,7 @@
             var configuration = new TelemetryConfiguration();
             new TestableTelemetryConfigurationFactory().Initialize(configuration, null, configFileContents);
 
-            Assert.Equal(3, configuration.TelemetryInitializers.Count); // Time, RoleInstance and operation initializers are added by default
+            Assert.Equal(2, configuration.TelemetryInitializers.Count); // Time and operation initializers are added by default
             Assert.NotNull(configuration.TelemetryInitializers.First(item => item.GetType().Name == "StubTelemetryInitializer"));
         }